Reorganize dk.brics code into root src tree
This is a *major* reogranization of code and structure, though, it is
done as to better integrate the two components of dk.brics together.
That is, `dk.brics.string` depends on `dk.brics.automaton`.  The old
structure did not facilitate building these components together.  This
new structure, on the otherhand, does.

Signed-off-by: Kenny Ballou <kballou@devnulllabs.io>